DB introduced - api endpoint added
