+++ database/models/blogs.js
@@ -1,0 +1,7 @@
+const mongoose = require('mongoose')
+
+const blogSchema = new mongoose.Schema({
+    url: String
+}, { timestamps: true })
+
+module.exports = mongoose.model('Blog', blogSchema)
